[ re #9 ] Basic diagram visualiser based on a circuit's ANF (#10)
* [ new ] start of a dot generator

* [ new ] Elaboration to A normal form

Looks like it will be easier to deal with simplified gates.
We expect something like:

foo(A,B,C) = A&B,Z where
  Z = A|B&C

to be elaborated to

foo(A,B,C) = RHS,Z where
  RHS   = A & B
  Z     = A | RIGHT
  RIGHT = B & C

i.e. only variables as outputs and equations are only using expressions
at most one gate deep.

Not quite finished yet.

* [ done ] Elaboration to A normal forms

* [ test ] for Anf

* [ show ] nicer instance for Def

* [ cleanup ] toANF

That being said, Dot should work directly on Gate rather than Def.

* [ test ] nesting on both sides.

* [ doc ] for the Anf pass

* [ fix ] remove undefined

* [ refactor ] Dot to use Anf

* [ refactor ] inputs and outputs get different names

As explained in the comment, when we are drawing the graph it is
*really* convenient to be sure that even if an input is outputted
immediately, the output got a different name.

e.g.

swap (X,Y) = Y,X

is hard to draw. But the equivalent

swap (X,Y) = A,B where
  A = Y
  B = X

is vastly simpler to handle.

* [ dot ] whitebox gate

* [ cleanup ] forget about aliases

Instead of outputting the wires as we discover them, we collect
them all. We then proceed to shrink all the paths of the form
output -> virtual node -> input.

* [ test ] move sample definitions to Smp

* [ fix ] add splitting node for shared wire

When a wire is used by more than 1 gate as an input it is nicer to
have an explicit point where it is splitted.

Note: does not work for outputs yet (so the tff example is broken)

* [ cleanup ] comments + abstract away declareAlias

* [ cleanup ] some more comments cleanup

* [ refactor ] extract MonadFresh properly

* [ refactor ] better representation of graphs

Next step: retain the left-to-right constraints and make sure that
wire shrinking is done in a left-to-right constraint preserving
manner.

* [ refactor ] complete overall of the circuit generation

Now using dot's record syntax to make sure the ports are kept in
the right order. Had to give up on straight wire lines as ports
are currently not supported with "splines = ortho;".

* [ test ] new example: rca4

* [ fix ] drop the nodes removed by shrinking

* [ test ] weird example